chore: fix #30
