--- src/core/parser/bool/structure.rs
@@ -1,5 +1,0 @@
-#[derive(PartialEq, Debug)]
-pub enum Bool {
-    True,
-    False,
-}
