+++ src/core/parser/equals.rs
@@ -1,0 +1,5 @@
+use nom::{character::complete::char, IResult};
+
+pub fn equals(input: &str) -> IResult<&str, char> {
+    char('=')(input)
+}
