--- src/core/parser/identifier/structure.rs
@@ -1,4 +1,0 @@
-#[derive(Debug, PartialEq)]
-pub struct Identifier {
-    pub value: String,
-}
