--- src/core/parser/list/structure.rs
@@ -1,6 +1,0 @@
-use crate::core::parser::bool::Bool;
-
-#[derive(Debug, PartialEq)]
-pub struct List {
-    pub values: Vec<Bool>,
-}
