--- src/core/parser/prelude/result.rs
@@ -1,4 +1,0 @@
-use nom::error::Error;
-use nom::IResult;
-
-pub type Result<T, U> = IResult<T, U, Error<T>>;
