+++ src/core/parser/underscore.rs
@@ -1,0 +1,5 @@
+use nom::{character::complete::char, IResult};
+
+pub fn underscore(input: &str) -> IResult<&str, char> {
+    char('_')(input)
+}
