--- src/core/parser/variable/structure.rs
@@ -1,8 +1,0 @@
-use crate::core::parser::bool::Bool;
-use crate::core::parser::identifier::Identifier;
-
-#[derive(Debug, PartialEq)]
-pub struct Variable {
-    pub identifier: Identifier,
-    pub value: Bool,
-}
